+++ forge/changelog.md
@@ -1,1 +1,2 @@
 - Added Esperanto translation
+- Fall back to Java microphone implementation when using ALSoft >= 1.25.0